Change to method hooks instead of metaprogramming.
While the metaprogramming method looked very clean, it did not allow for
the use of helper methods. This was due to the value of self in the
block being the Class object, not the instance object.

It was considered to use `self.instance_eval &block` but this did not
allow for the response data to be passed into the block.